Alterando long para integer no parametro idProposicao pensandoodireito/sislegis-app#170

+++ src/main/java/br/gov/mj/sislegis/app/parser/senado/ParserVotacaoSenado.java
@@ -14,7 +14,7 @@ import java.util.Locale;
 
 public class ParserVotacaoSenado {
 
-    public List<Votacao> votacoesPorProposicao(Long idProposicao) throws Exception {
+    public List<Votacao> votacoesPorProposicao(Integer idProposicao) throws Exception {
         List<Votacao> votacoes = new ArrayList<>();
         StringBuilder wsURL = new StringBuilder("http://legis.senado.leg.br/dadosabertos/materia/votacoes/").append(idProposicao);
